# Inline the `isString` check in the `Parser.getObj` method